declarationsが読み込めなかったので修正
