+++ tsconfig.json
@@ -1,5 +1,7 @@
 {
   "compilerOptions": {
+    "strict": true,
+    "esModuleInterop": true,
     "module": "commonjs"
   }
 }
